📌 fix: 본인 AnswerPage에서 삭제하기 버튼 안보이는 에러 해결
