+++ src/pages/AnswerPage/AnswerPage.style.jsx
@@ -104,6 +104,37 @@ const DeleteButtonDiv = styled.div`
 // 몇 개의 질문이 있습니다 div
 const QuestionDiv = styled.div``;
 
+const FloatingDeleteButton = styled.div`
+  position: absolute;
+  top: 35.5rem;
+  right: 41rem;
+  cursor: pointer;
+  display: flex;
+  width: 100px;
+  height: 35px;
+  padding: 12px 24px;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+  flex-shrink: 0;
+
+  border-radius: 200px;
+  background: var(--brown-40, #542f1a);
+
+  /* 3pt */
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+
+  p {
+    color: var(--grayscale-10, #fff);
+    font-feature-settings: 'clig' off, 'liga' off;
+    font-family: Actor;
+    font-size: 15px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 25px; /* 166.667% */
+  }
+`;
+
 export {
   HeaderImage,
   HeaderDiv,
@@ -116,4 +147,5 @@ export {
   QuestionDiv,
   DeleteButtonDiv,
   QuestionContainer,
+  FloatingDeleteButton,
 };
