+++ src/pages/FeedPage/FeedPage.jsx
@@ -76,7 +76,7 @@ const FeedPage = () => {
           <List questions={questions} subjectData={subjectData} />
         )}
 
-        {/* Subject를 작성한 것이 User인지 확인하는 함수: User일 시 질문 작성하기 버튼 안보이게 해야 함  checkUser(subjectId) 여야 함 */}
+        {/* Subject를 작성한 것이 User인지 확인하는 함수: User일 시 질문 작성하기 버튼 안보이게 해야 함  */}
         {!checkUser(subjectId) && (
           <S.ButtonWrapper>
             <FloatingButton setActive={setActive} />
