+++ src/utils/checkUser.js
@@ -2,5 +2,5 @@ import { USER_ID } from '../constants/constant';
 
 // 해당 Subject를 생성한 유저인지 확인
 export const checkUser = (subjectId) => {
-  return USER_ID !== subjectId;
+  return USER_ID === subjectId;
 };
